4.21

+++ osa_4/blogilista/controllers/blogs.js
@@ -33,6 +33,18 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+  console.log(blogToDelete.user.toString())
+  if (blogToDelete.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'can not delete other users blogs'})
+  }
+  
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
